Ice and Temps. Working - Added experimental rotation stuff in activeGraphics.
Ice pressure and temperatures are now completely working. Have worked
on splitting up the graphics and ordering to that arcs for pressure
don't interfere with static graphics. Have worked on how and where we
would rotate the temperatures and ice pressures. Have added a little
code for this in the activeGraphics method, however this isn't 100%
reliable, as I had to manually translate (reposition) where the centre
of the graphics was. Need to find a way to separate temps and pressure
together so we can rotate them free of the other graphics.